[V3] (part 1) Change runner/v3 to runner/v2. Update v3 flag to v2
