+++ cmd/skaffold/app/cmd/flags.go
@@ -528,8 +528,8 @@ var flagRegistry = []Flag{
 		DefinedOn:     []string{"dev", "build", "run", "debug", "deploy"},
 	},
 	{
-		Name:          "v3",
-		Usage:         "Next skaffold config (v3). Use kpt to render/hydrate and deploy manifests.",
+		Name:          "v2",
+		Usage:         "Next skaffold config (v2). Use kpt to render/hydrate and deploy manifests.",
 		Value:         &opts.Experimental,
 		DefValue:      false,
 		FlagAddMethod: "BoolVar",
