+++ pkg/skaffold/runner/v2/runner.go
@@ -13,16 +13,16 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package v3
+package v2
 
 import (
-	"context"
-	"fmt"
-	"io"
-
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/graph"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/runner"
 )
 
-func (r *SkaffoldRunner) Render(ctx context.Context, out io.Writer, builds []graph.Artifact, offline bool, filepath string) error {
-	return fmt.Errorf("not implemented error: SkaffoldRunner(v3).Dev")
+type SkaffoldRunner struct {
+	runner.Builder
+	runner.Pruner
+	runner.Tester
 }
+
+func (r *SkaffoldRunner) HasDeployed() bool { return true }
